refactor: extract core package
